+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Added `core::error::Error` implementations to all error types.
 - `SystemTable::exit_boot_services` now takes one param `memory_type` to ensure
   the memory type of memory map.
+- Added the `ShellParams` protocol
 
 ### Removed
 - `BootServices::memmove` and `BootServices::set_mem` have been removed, use